fix: localstore error on vite
