+++ src/components/Apps/ChatAi/index.vue
@@ -51,7 +51,7 @@ import Sidebar from './components/Sidebar/index.vue'
 import Chat from './components/ChitChat/index.vue'
 import { getInputFocus } from './useChat.js'
 import DrawerPanel from '@/components/Apps/DrawerPanel/index.vue'
-import { ObjectLocalStorage } from '@/utils/common'
+import { ObjectLocalStorage } from '@/utils/common/objectLocalStorage'
 import i18n from '@/i18n/i18n'
 import { getAssetUrl } from '@/utils/assets'
 import { mapGetters } from 'vuex'
