+++ src/layout/components/NavHeader/Search.vue
@@ -131,7 +131,7 @@
 
 <script>
 import Icon from '@/components/Widgets/Icon/index.vue'
-import { ObjectLocalStorage } from '@/utils/common'
+import { ObjectLocalStorage } from '@/utils/common/objectLocalStorage'
 import { Search as SearchIcon } from '@element-plus/icons-vue'
 import _ from 'lodash'
 import { markRaw } from 'vue'
